# docs/openapi/declaration.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RÉCOR Declaration Service",
    "description": "REST surface of the RÉCOR Declaration service. Accepts beneficial-ownership declarations, returns signed receipts, relays verification outcomes from the Verification Engine, and exposes operator endpoints under `/v1/internal/*`. Endpoints under the `internal` tag are not for general consumer use.",
    "contact": {
      "name": "RÉCOR domain team",
      "email": "noreply@recor.cm"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local dev"
    },
    {
      "url": "https://api.recor.cm",
      "description": "Production (TBD)"
    }
  ],
  "paths": {
...
    "/v1/internal/outbox-dlq": {
      "get": {
        "tags": [
          "internal"
        ],
        "operationId": "listDlq",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of rows to return; defaults to 50.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 50
            },
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Zero-indexed offset for paging.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            },
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Page of DLQ rows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDlqResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Principal not on the admin allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Backend failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Admin endpoints disabled (allowlist empty)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "devPrincipalHeader": []
          }
        ]
      }
    },
    "/v1/internal/outbox-dlq/{id}/replay": {
      "post": {
        "tags": [
          "internal"
        ],
        "operationId": "replayDlq",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "DLQ row id to replay",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Row was moved back to the outbox for retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayDlqResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Principal not on the admin allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "No DLQ row with the given id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Backend failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Admin endpoints disabled (allowlist empty)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "devPrincipalHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DlqItem": {
        "type": "object",
        "required": [
          "id",
          "event_id",
          "event_type",
          "event_version",
          "aggregate_type",
          "aggregate_id",
          "partition_key",
          "payload",
          "created_at",
          "dead_lettered_at",
          "dispatch_attempts"
        ],
        "properties": {
          "aggregate_id": {
            "type": "string",
            "format": "uuid"
          },
          "aggregate_type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "dead_lettered_at": {
            "type": "string",
            "format": "date-time"
          },
          "dispatch_attempts": {
            "type": "integer",
            "format": "int32"
          },
          "event_id": {
            "type": "string",
            "format": "uuid"
          },
          "event_type": {
            "type": "string"
          },
          "event_version": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "partition_key": {
            "type": "string"
          },
          "payload": {
            "type": "object",
            "description": "Original event payload as JSON. Schema depends on `event_type`."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Inner body of an `ErrorEnvelope`. `kind` is a stable machine-friendly\nclassifier — clients SHOULD switch on it. `message` is a human\ndescription for logs; clients SHOULD NOT parse it.",
        "required": [
          "kind",
          "message"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Machine-friendly classifier (e.g. `bad_request`, `not_found`,\n`conflict`, `forbidden`, `authentication_required`,\n`idempotency_conflict`, `bad_attestation`, `internal`,\n`optimistic_concurrency_conflict`, `admin_disabled`, `not_admin`,\n`missing_signature`, `bad_signature`, `writeback_disabled`,\n`dlq_row_not_found`, `malformed_envelope`, `malformed_payload`)."
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "description": "Standard error response body. Every non-2xx response is shaped like\n`{ \"error\": { \"kind\": \"<machine>\", \"message\": \"<human>\" } }`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "ListDlqResponse": {
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DlqItem"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "description": "Limit echoed back from the query string."
          },
          "offset": {
            "type": "integer",
            "format": "int64",
            "description": "Offset echoed back from the query string."
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of DLQ rows currently held (unfiltered)."
          }
        }
      },
...
      "ReplayDlqResponse": {
        "type": "object",
        "required": [
          "id",
          "replayed"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "replayed": {
            "type": "boolean",
            "description": "Always `true` on success; on failure the response is an error\nenvelope, not this struct."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC-issued JWT. The service verifies the signature against the configured issuer's JWKS, plus `iss`/`aud`/`exp`/`nbf` claims. HMAC algorithms are refused (algorithm-confusion mitigation)."
      },
      "devPrincipalHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Recor-Dev-Principal",
        "description": "Dev-only principal shortcut. Refused outside ENVIRONMENT=dev. Production must always use `bearerAuth`."
      },
      "hmacSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RECOR-Signature",
        "description": "HMAC-SHA256 over the raw request body, hex-encoded, using the shared writeback secret. Constant-time verification; rotation supported via a transient old-secret window."
      }
    }
  },
...
}